+++ src/core/document/spatial_query.rs
@@ -835,7 +835,7 @@ where
   V: SpatialVisitor,
 {
   fn grow(&mut self, count: usize) -> Result<()> {
-    self.result.grow(count as i32)?;
+    self.result.grow(count)?;
     Ok(())
   }
