+++ src/core/util/packed/bulk_operation_packed1.rs
@@ -32,7 +32,7 @@ impl Decoder for BulkOperationPacked1 {
     mut blocks_offset: usize,
     values: &mut [i64],
     mut values_offset: usize,
-    iterations: i32,
+    iterations: usize,
   ) {
     for _ in 0..iterations {
       let block = blocks[blocks_offset];
@@ -52,7 +52,7 @@ impl Decoder for BulkOperationPacked1 {
     mut blocks_offset: usize,
     values: &mut [i64],
     mut values_offset: usize,
-    iterations: i32,
+    iterations: usize,
   ) {
     for _ in 0..iterations {
       let block = blocks[blocks_offset];
@@ -78,7 +78,7 @@ impl Decoder for BulkOperationPacked1 {
     mut blocks_offset: usize,
     values: &mut [i32],
     mut values_offset: usize,
-    iterations: i32,
+    iterations: usize,
   ) -> Result<()> {
     for _ in 0..iterations {
       let block = blocks[blocks_offset];
@@ -99,7 +99,7 @@ impl Decoder for BulkOperationPacked1 {
     mut blocks_offset: usize,
     values: &mut [i32],
     mut values_offset: usize,
-    iterations: i32,
+    iterations: usize,
   ) -> Result<()> {
     for _ in 0..iterations {
       let block = blocks[blocks_offset];
