+++ src/core/util/packed/bulk_operation_packed16.rs
@@ -31,7 +31,7 @@ impl Decoder for BulkOperationPacked16 {
     mut blocks_offset: usize,
     values: &mut [i64],
     mut values_offset: usize,
-    iterations: i32,
+    iterations: usize,
   ) {
     for _ in 0..iterations {
       let block = blocks[blocks_offset];
@@ -49,7 +49,7 @@ impl Decoder for BulkOperationPacked16 {
     mut blocks_offset: usize,
     values: &mut [i64],
     mut values_offset: usize,
-    iterations: i32,
+    iterations: usize,
   ) {
     for _ in 0..iterations {
       values[values_offset] =
@@ -65,7 +65,7 @@ impl Decoder for BulkOperationPacked16 {
     mut blocks_offset: usize,
     values: &mut [i32],
     mut values_offset: usize,
-    iterations: i32,
+    iterations: usize,
   ) -> Result<()> {
     for _ in 0..iterations {
       let block = blocks[blocks_offset];
@@ -84,7 +84,7 @@ impl Decoder for BulkOperationPacked16 {
     mut blocks_offset: usize,
     values: &mut [i32],
     mut values_offset: usize,
-    iterations: i32,
+    iterations: usize,
   ) -> Result<()> {
     for _ in 0..iterations {
       values[values_offset] =
